just need to create txt file
In the process of working on this code, I learned that I had a very weak grasp on scope in loops. By basing my code off the general structure of code from Saturday's classwork, I managed to accidentally set up two loops, which prevented my file from ever functioning as expected. Once I was able to identify and fix that issue, the rest of the code fell into place fairly easily. Once I add the code that will cause this file to create a new text file featuring the requested summary table at the completion of the loop, I will be ready to turn this in for grading.